roachprod: add node_exporter and collect on AWS and Azure
Previously, Prometheus was only collecting the `cockroachdb` metrics on
the GCP roachprod clusters.

This PR introduces the installation and collection of system metrics
exposed via `node_exporter` now installed during in the nodes' startup
script, and also adds metrics collection for AWS and Azure.

The difference between secure and insecure targets is now managed by
Prometheus labels, which allows to scrape both secure and insecure
exporters on the same cluster.

Firewall rules are added to prevent node_exporter metrics collection
from untrusted sources.
The node_exporter setup inside the `grafana-start` command has been
reworked a bit to avoid reinstalling node_exporter if it is already
installed on the node, and to adapt the firewall rules so that it can be
scrapped by the cluster's Prometheus node.

Epic: none
Release note: None

+++ pkg/roachprod/prometheus/prometheus.go
@@ -11,6 +11,7 @@ import (
 	"os"
 	"path/filepath"
 	"strconv"
+	"strings"
 	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/roachprod/install"
@@ -22,7 +23,10 @@ import (
 	"gopkg.in/yaml.v2"
 )
 
-const defaultWorkloadPort = 2112
+const (
+	defaultWorkloadPort = 2112
+	nodeExporterStarted = "node_exporter/.started_with_start_grafana"
+)
 
 // Client is an interface allowing queries against Prometheus.
 type Client interface {
@@ -169,16 +173,24 @@ func (cfg *Config) WithScrapeConfigs(config ...ScrapeConfig) *Config {
 // For more on the node exporter process, see https://prometheus.io/docs/guides/node-exporter/
 func (cfg *Config) WithNodeExporter(nodes install.Nodes) *Config {
 	cfg.NodeExporter = nodes
+
+	port, err := strconv.Atoi(vm.NodeExporterPort)
+	if err != nil {
+		// If the port is not a number, default to 9100.
+		// This should never happen unless the constant is changed.
+		port = 9100
+	}
 	// Add a scrape config for each node running node_exporter
 	for _, node := range cfg.NodeExporter {
 		s := strconv.Itoa(int(node))
 		cfg.ScrapeConfigs = append(cfg.ScrapeConfigs, ScrapeConfig{
 			JobName:     "node_exporter-" + s,
-			MetricsPath: "/metrics",
+			MetricsPath: vm.NodeExporterMetricsPath,
 			Labels:      map[string]string{"node": s},
 			ScrapeNodes: []ScrapeNode{{
 				Node: node,
-				Port: 9100}},
+				Port: port,
+			}},
 		})
 	}
 	return cfg
@@ -237,11 +249,16 @@ type Prometheus struct {
 
 // Init creates a prometheus instance on the given cluster.
 func Init(
-	ctx context.Context, l *logger.Logger, c *install.SyncedCluster, arch vm.CPUArch, cfg Config,
+	ctx context.Context, l *logger.Logger, c *install.SyncedCluster, cfg Config,
 ) (_ *Prometheus, _ error) {
-	binArch := "amd64"
-	if arch == vm.ArchARM64 {
-		binArch = "arm64"
+
+	nodeIPs, err := makeNodeIPMap(c)
+	if err != nil {
+		return nil, err
+	}
+	prometheusNodeIPs := make([]string, 0, len(cfg.PrometheusNode))
+	for _, node := range cfg.PrometheusNode {
+		prometheusNodeIPs = append(prometheusNodeIPs, nodeIPs[node])
 	}
 
 	if len(cfg.NodeExporter) > 0 {
@@ -251,18 +268,31 @@ func Init(
 		if err := c.Run(ctx, l, l.Stdout, l.Stderr, install.WithNodes(cfg.NodeExporter).WithShouldRetryFn(install.AlwaysTrue),
 			"download node exporter",
 			fmt.Sprintf(`
-(sudo systemctl stop node_exporter || true) &&
-rm -rf node_exporter && mkdir -p node_exporter && curl -fsSL \
-  https://storage.googleapis.com/cockroach-test-artifacts/prometheus/node_exporter-1.2.2.linux-%s.tar.gz |
-  tar zxv --strip-components 1 -C node_exporter
-`, binArch)); err != nil {
+export ARCH=$(dpkg --print-architecture)
+if ! $(systemctl is-active --quiet node_exporter); then
+	(sudo systemctl stop node_exporter || true) &&
+	rm -rf node_exporter && mkdir -p node_exporter && curl -fsSL \
+		https://storage.googleapis.com/cockroach-test-artifacts/prometheus/node_exporter-%s.linux-${ARCH}.tar.gz |
+		tar zxv --strip-components 1 -C node_exporter
+fi
+`,
+				vm.NodeExporterVersion,
+			)); err != nil {
 			return nil, err
 		}
 
 		// Start node_exporter.
+		// Also make it reachable by the Prometheus node in case of firewall rules.
 		if err := c.Run(ctx, l, l.Stdout, l.Stderr, install.WithNodes(cfg.NodeExporter), "init node exporter",
-			`cd node_exporter &&
-sudo systemd-run --unit node_exporter --same-dir ./node_exporter`,
+			fmt.Sprintf(`
+sudo iptables -C INPUT -p tcp -s %[1]s --dport %[2]s -j ACCEPT || sudo iptables -I INPUT -p tcp -s %[1]s --dport %[2]s -j ACCEPT;
+if ! $(systemctl is-active --quiet node_exporter); then
+	touch %[4]s &&
+	cd node_exporter &&
+	sudo systemd-run --unit node_exporter --same-dir ./node_exporter \
+		--web.listen-address=":%[2]s" --web.telemetry-path="%[3]s"
+fi
+`, strings.Join(prometheusNodeIPs, ","), vm.NodeExporterPort, vm.NodeExporterMetricsPath, nodeExporterStarted),
 		); err != nil {
 			// TODO(msbutler): download binary for target platform. currently we
 			// hardcode downloading the linux binary.
@@ -288,16 +318,15 @@ sudo systemd-run --unit node_exporter --same-dir ./node_exporter`,
 		l.Stderr,
 		install.WithNodes(cfg.PrometheusNode).WithShouldRetryFn(install.AlwaysTrue),
 		"download prometheus",
-		fmt.Sprintf(`sudo rm -rf /tmp/prometheus && mkdir /tmp/prometheus && cd /tmp/prometheus &&
-			curl -fsSL https://storage.googleapis.com/cockroach-test-artifacts/prometheus/prometheus-2.27.1.linux-%s.tar.gz | tar zxv --strip-components=1`,
-			binArch)); err != nil {
+		fmt.Sprintf(`
+export ARCH=$(dpkg --print-architecture)
+sudo rm -rf /tmp/prometheus && mkdir /tmp/prometheus && cd /tmp/prometheus &&
+curl -fsSL https://storage.googleapis.com/cockroach-test-artifacts/prometheus/prometheus-%s.linux-${ARCH}.tar.gz | tar zxv --strip-components=1`,
+			vm.PrometheusVersion,
+		)); err != nil {
 		return nil, err
 	}
 	// create and upload prom config
-	nodeIPs, err := makeNodeIPMap(c)
-	if err != nil {
-		return nil, err
-	}
 	yamlCfg, err := makeYAMLConfig(cfg.ScrapeConfigs, nodeIPs)
 	if err != nil {
 		return nil, err
@@ -335,14 +364,16 @@ sudo systemd-run --unit prometheus --same-dir \
 			l.Stdout,
 			l.Stderr, install.WithNodes(cfg.PrometheusNode).WithShouldRetryFn(install.AlwaysTrue), "install grafana",
 			fmt.Sprintf(`
+export ARCH=$(dpkg --print-architecture)
 sudo apt-get install -qqy apt-transport-https &&
 sudo apt-get install -qqy software-properties-common &&
 sudo apt-get install -y adduser libfontconfig1 &&
-echo "Downloading https://dl.grafana.com/enterprise/release/grafana-enterprise_9.2.3_%[1]s.deb" &&
-curl https://dl.grafana.com/enterprise/release/grafana-enterprise_9.2.3_%[1]s.deb -sS -o grafana-enterprise_9.2.3_%[1]s.deb &&
-sudo dpkg -i grafana-enterprise_9.2.3_%[1]s.deb &&
+echo "Downloading https://dl.grafana.com/enterprise/release/grafana-enterprise_%[1]s_${ARCH}.deb" &&
+curl https://dl.grafana.com/enterprise/release/grafana-enterprise_%[1]s_${ARCH}.deb -sS -o grafana-enterprise_%[1]s_${ARCH}.deb &&
+sudo dpkg -i grafana-enterprise_%[1]s_${ARCH}.deb &&
 sudo mkdir -p /var/lib/grafana/dashboards`,
-				binArch)); err != nil {
+				vm.GrafanaEnterpriseVersion,
+			)); err != nil {
 			return nil, err
 		}
 
@@ -484,6 +515,12 @@ func Shutdown(
 	// We currently assume the last node contains the server.
 	promNode := install.Nodes{nodes[len(nodes)-1]}
 
+	nodeIPs, err := makeNodeIPMap(c)
+	if err != nil {
+		return err
+	}
+	promIP := nodeIPs[promNode[0]]
+
 	dumpSnapshot := func(dumpDir string) error {
 		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
 		defer cancel()
@@ -499,8 +536,15 @@ func Shutdown(
 			shutdownErr = errors.CombineErrors(shutdownErr, err)
 		}
 	}
+	// Stop node_exporter if it was started by grafana-start.
+	// Drop the firewall rule added for the Prometheus node.
 	if err := c.Run(ctx, l, l.Stdout, l.Stderr, install.WithNodes(nodes), "stop node exporter",
-		`sudo systemctl stop node_exporter || echo 'Stopped node exporter'`); err != nil {
+		fmt.Sprintf(`
+sudo iptables -D INPUT -p tcp -s %[1]s --dport %[2]s -j ACCEPT || true;
+if [ -f %[3]s ] ; then
+	sudo systemctl stop node_exporter || echo 'Stopped node exporter'
+fi
+`, promIP, vm.NodeExporterPort, nodeExporterStarted)); err != nil {
 		l.Printf("Failed to stop node exporter: %v", err)
 		shutdownErr = errors.CombineErrors(shutdownErr, err)
 	}
